[TPTP] Rename plugin + fix * fix dot end of line * fix cut * add letter before conjecture * rename package
